+++ backend/company_app/views.py
@@ -4,6 +4,7 @@ from rest_framework import status
 from django.shortcuts import get_object_or_404
 from .models import Company
 from .serializers import CompanySerializer
+
 class CompanyView(APIView):
     
     def get(self, request, pk=None):
@@ -38,6 +39,10 @@ class CompanyView(APIView):
         company = get_object_or_404(Company, id=pk)
         hard = request.query_params.get("hard", "false").lower() == "true"
         if hard:
+            if request.user.role != "admin":
+                raise self.permission_denied(request,
+                message="Solo admin puede realizar borrado físico."
+            );
             company.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)    
         company.enabled = False
